Move App::setCurrentLanguage and fix type hinting
