+++ src/Cms/App.php
@@ -31,6 +31,7 @@ use Kirby\Toolkit\Config;
 use Kirby\Toolkit\Controller;
 use Kirby\Toolkit\LazyValue;
 use Kirby\Toolkit\Str;
+use Kirby\Toolkit\Locale;
 use Kirby\Uuid\Uuid;
 use Throwable;
 
@@ -1213,7 +1214,7 @@ class App
 	 * @internal
 	 * @throws \Kirby\Exception\NotFoundException if the home page cannot be found
 	 */
-	public function resolve(string $path = null, string $language = null): mixed
+	public function resolve(string|null $path = null, string|null $language = null): mixed
 	{
 		// set the current translation
 		$this->setCurrentTranslation($language);
@@ -1409,6 +1410,30 @@ class App
 			$this->root('sessions'),
 			$this->option('session', [])
 		);
+	}
+
+	/**
+	 * Load and set the current language if it exists
+	 * Otherwise fall back to the default language
+	 *
+	 * @internal
+	 */
+	public function setCurrentLanguage(
+		string|null $languageCode = null
+	): Language|null {
+		if ($this->multilang() === false) {
+			Locale::set($this->option('locale', 'en_US.utf-8'));
+			return $this->language = null;
+		}
+
+		$this->language = $this->language($languageCode) ?? $this->defaultLanguage();
+
+		Locale::set($this->language->locale());
+
+		// add language slug rules to Str class
+		Str::$language = $this->language->rules();
+
+		return $this->language;
 	}
 
 	/**
